I've rewritten track.js on it's entirety. It's a paradigm shift from what we had before.
Now:

Track -> holds name, description, trackpoints and averages (precomputed). It can be sliced, bounds can be found and nearest too, as before.

Trackpoint -> holds all instantaneous stats. I've done this since it is way easier to manage data back and forth with D3 and the datum option if everything I want to represent is precomputed into an array, instead of having to to the calculus manually in render or some other place.

parseTrack -> Async like function. This will be the function used when instantiating a track from an XML. Since it's async, it won't block the browser, as it did before. Being non blocking is important, since there are (potentially) lots of calculations to perform.

Also, as things are written now, they will take advantage of hidden classes in V8, so it should be snappy performance wise (many improvements pending, I suppose).

Anyway, even though it's a rehash of what was already working, I have not tested it, so it is probably broken in quite a lot of ways. In further commits I will get the tests up and running again and perform the necessary refactors to fix this branch to a working status